deduplicate config & local-storage
