+++ src/components/menus/directory-tree-contextual-menu.jsx
@@ -24,11 +24,10 @@ import DeleteDialog from '../dialogs/delete-dialog';
 import { DialogsId } from '../../utils/UIconstants';
 import CommonContextualMenu from './common-contextual-menu';
 import { useDeferredFetch } from '../../utils/custom-hooks';
-import { ElementType, FilterCreationDialog, useSnackMessage } from '@gridsuite/commons-ui';
+import { ElementType, FilterCreationDialog, PARAM_LANGUAGE, useSnackMessage } from '@gridsuite/commons-ui';
 import ContingencyListCreationDialog from '../dialogs/contingency-list/creation/contingency-list-creation-dialog';
 import CreateCaseDialog from '../dialogs/create-case-dialog/create-case-dialog';
 import { useParameterState } from '../dialogs/parameters-dialog';
-import { PARAM_LANGUAGE } from '../../utils/config-params';
 import { handleMaxElementsExceededError } from '../utils/rest-errors';
 import { directorySrv, exploreSrv } from '../../services';
