--- src/utils/config-params.js
@@ -1,18 +1,0 @@
-/**
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-export const COMMON_APP_NAME = 'common';
-export const APP_NAME = 'Explore';
-
-export const PARAM_THEME = 'theme';
-export const PARAM_LANGUAGE = 'language';
-
-const COMMON_CONFIG_PARAMS_NAMES = new Set([PARAM_THEME, PARAM_LANGUAGE]);
-
-export function getAppName(paramName) {
-    return COMMON_CONFIG_PARAMS_NAMES.has(paramName) ? COMMON_APP_NAME : APP_NAME;
-}
