+++ src/utils/config-params.ts
@@ -1,12 +1,8 @@
 /**
- * Copyright (c) 2024, RTE (http://www.rte-france.com)
+ * Copyright (c) 2021, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-export const noSelectionForCopy = {
-    sourceCaseUuid: null,
-    name: null,
-    description: null,
-    parentDirectoryUuid: null,
-};
+
+export const APP_NAME = 'Explore';
